VBScripts for Microsoft Word
VBScripts for Microsoft Word 
Create Word Combinations 
This script will create a list of all of the 4-letter combinations within a Word document. This is strictly a  learning exercise, and has no other value. 
Note: There is code in the program to stop processing after a set number of words are created. If you let the  entire program process, it will (try to) create almost a half million lines. 
Features: 
∙ Add lines to a Microsoft Word document 
∙ Create a line set to a specific font name and font size